+++ administrator/components/com_loginguard/src/Model/AttemptsModel.php
@@ -13,8 +13,7 @@ final class AttemptsModel extends ListModel
     {
         if (empty($config['filter_fields'])) {
             $config['filter_fields'] = [
-                'id', 'ip_address', 'name', 'username', 'status', 'created', 'reason', 'country', 'country_code',
-                'region', 'city', 'isp', 'asn', 'browser', 'operating_system', 'where_at', 'client', 'user_agent',
+                'id', 'ip_address', 'name', 'username', 'status', 'created', 'reason', 'browser', 'operating_system', 'where_at', 'client', 'user_agent',
                 'attempt_type', 'mfa_method',
             ];
         }
@@ -73,9 +72,7 @@ final class AttemptsModel extends ListModel
         $query = $db->getQuery(true)
             ->select([
                 $db->quoteName('id'), $db->quoteName('ip_address'), $db->quoteName('name'), $db->quoteName('username'),
-                $db->quoteName('status'), $db->quoteName('created'), $db->quoteName('reason'), $db->quoteName('country'),
-                $db->quoteName('country_code'), $db->quoteName('region'), $db->quoteName('city'), $db->quoteName('isp'),
-                $db->quoteName('asn'), $db->quoteName('browser'), $db->quoteName('operating_system'), $db->quoteName('where_at'),
+                $db->quoteName('status'), $db->quoteName('created'), $db->quoteName('reason'), $db->quoteName('browser'), $db->quoteName('operating_system'), $db->quoteName('where_at'),
                 $db->quoteName('client'), $db->quoteName('user_agent'), $db->quoteName('attempt_type'), $db->quoteName('mfa_method'),
             ])
             ->from($db->quoteName('#__loginguard_attempts'));
@@ -84,7 +81,7 @@ final class AttemptsModel extends ListModel
         if ($search !== '') {
             $pattern = '%' . str_replace(' ', '%', $search) . '%';
             $fields = [
-                'ip_address', 'name', 'username', 'country', 'country_code', 'region', 'city', 'isp', 'asn', 'browser',
+                'ip_address', 'name', 'username', 'browser',
                 'operating_system', 'reason', 'user_agent', 'where_at', 'client', 'attempt_type', 'mfa_method',
             ];
             $conditions = array_map(static fn ($field) => $db->quoteName($field) . ' LIKE ' . $db->quote($pattern), $fields);